Convert wpa_supplicant related module to soong
The conversion skip those condition which not belong to Android to
simplify the conversion, you can reference aosp/3218836 for those unused
if condition in Android.mk which is not put in the CL to preveninting
merge conflict. Also skip hostapd_noaidl due to seems there's no other
place reference it.

Bug: 325571506
Test: m hostapd_cli hostapd libhostapd_aidl wpa_cli wpa_supplicant \
        wpa_supplicant_macsec libwpa_client libwpa_aidl libpasn
Flag: EXEMPT refactor

Change-Id: Ic41de8907377144b9aa94322bbfa3ae2667649c9